src/apps/requirements.txt
a2wsgi==1.10.8
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
blinker==1.6.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.2
connexion==3.2.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-swagger-ui==4.11.1
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.1
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg==3.2.3
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.22.3
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
starlette==0.45.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
zipp==3.21.0
